chore: refactor swap indexing
