+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { keccak256 } from 'viem';
+import { keccak256, formatUnits } from 'viem';
 import { moduleTitles } from './constants';
 
 export const getModuleTitleFromId = (hash: `0x${string}`) => {
@@ -6,4 +6,8 @@ export const getModuleTitleFromId = (hash: `0x${string}`) => {
     // @ts-ignore
     return keccak256(title) === hash;
   });
+};
+
+export const uintToFloat = (amount: bigint, decimals: number) => {
+  return parseFloat(formatUnits(amount, decimals));
 };
